requirements.txt
# LlamaIndex & LLM related
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-ollama
chromadb
ollama # If you install the python client

# Watcher & Extraction related
watchdog
python-docx
pdfplumber
sentence-transformers # Needed by LlamaIndex's HuggingFaceEmbedding
scikit-learn # Needed by Project 2's fallback vectorstore (though we won't use it)
joblib # Needed by Project 2's fallback vectorstore (though we won't use it)

# General Python
pydantic # If using models.py
transformers
torch # Usually needed for sentence-transformers/HuggingFaceEmbedding
python-dotenv # Good practice for managing API keys if needed later